Add ArchUnit test for checking dependencies of the Declarative evaluator
