+++ platforms/core-configuration/declarative-dsl-provider/src/main/kotlin/org/gradle/internal/declarativedsl/interpreter/GradleProcessInterpretationSchemaBuilder.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.declarativedsl.evaluator
+package org.gradle.internal.declarativedsl.interpreter
 
 import org.gradle.api.internal.SettingsInternal
 import org.gradle.internal.declarativedsl.evaluator.schema.DeclarativeScriptContext
