+++ platforms/core-configuration/declarative-dsl-provider/src/main/kotlin/org/gradle/internal/declarativedsl/provider/DeclarativeDslServices.kt
@@ -20,12 +20,12 @@ import org.gradle.api.internal.GradleInternal
 import org.gradle.api.model.ObjectFactory
 import org.gradle.initialization.layout.BuildLayoutConfiguration
 import org.gradle.initialization.layout.BuildLayoutFactory
-import org.gradle.internal.declarativedsl.evaluator.DeclarativeKotlinScriptEvaluator
-import org.gradle.internal.declarativedsl.evaluator.GradleProcessInterpretationSchemaBuilder
-import org.gradle.internal.declarativedsl.evaluator.MemoizedInterpretationSchemaBuilder
-import org.gradle.internal.declarativedsl.evaluator.StoringInterpretationSchemaBuilder
-import org.gradle.internal.declarativedsl.evaluator.defaultDeclarativeScriptEvaluator
-import org.gradle.internal.declarativedsl.evaluator.defaults.DeclarativeModelDefaultsHandler
+import org.gradle.internal.declarativedsl.interpreter.DeclarativeKotlinScriptEvaluator
+import org.gradle.internal.declarativedsl.interpreter.GradleProcessInterpretationSchemaBuilder
+import org.gradle.internal.declarativedsl.interpreter.MemoizedInterpretationSchemaBuilder
+import org.gradle.internal.declarativedsl.interpreter.StoringInterpretationSchemaBuilder
+import org.gradle.internal.declarativedsl.interpreter.defaultDeclarativeScriptEvaluator
+import org.gradle.internal.declarativedsl.interpreter.defaults.DeclarativeModelDefaultsHandler
 import org.gradle.internal.declarativedsl.evaluator.schema.InterpretationSchemaBuilder
 import org.gradle.internal.event.ListenerManager
 import org.gradle.internal.service.Provides
