+++ platforms/core-configuration/declarative-dsl-tooling-models/build.gradle.kts
@@ -30,7 +30,7 @@ dependencies {
     compileOnly(libs.kotlinStdlib) {
         because(
             "used by the compiler, but there should be no binary dependency on the stdlib; " +
-                "this project should be usable by Kotlin-less consumers, see: `NoKotlinStdlibBinaryDependencyTest`"
+                "this project should be usable by Kotlin-less consumers, see: `DeclarativeDslDependencyTest`"
         )
     }
 }
